Re-enabled Gamepads
Prevented crashing if there are not enough gamepads by falling back to
keyboard
For some reason, if we do not use FlxG.gamepads.getActiveGamepads,
input doesn’t work, so I reassign Input.gamepads to
FlxG.gamepads.getActiveGamepads every frame in PlayState.

Added a start button to all players, which can be used as a pause key
later, and as a switching key for single input mode.