Adjusted Input handling
using FlxG.gamepads.getByID to return a nonexistent gamepad would make
the game unplayable if someone get the controls to gamepad and then
didn’t have a gamepad to set it back
uses getByID only once its sure that there is actually a gamepad, and
reverts to keyboard if there is no gamepad for the associated player.
Seems like a more robust way of doing it, having been trapped in poorly
coded stuff in the past where you set controls to controller and then
have a battery die or something and can’t change it to something else
until you get the original peripheral up and running again.